pyproject.toml
[project]
name = "chatbot"
version = "0.1.0"
description = "Streamlit chatbot with Anthropic integration"
authors = [
    {name = "Akshay Karle",email = "1443108+akshaykarle@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "streamlit>=1.32.0",
    "anthropic>=0.18.0",
    "python-dotenv>=1.0.0",
    "pytest>=7.4.0",
    "pytest-mock>=3.11.1",
    "flake8>=6.0.0",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
